README file updated and made some fixes.
Released On: 11/13/2024

I updated the README.md file to make it more descriptive. Instead of just the name and project description, it now includes the required software, how to install it, and how to use it. I added more comments to the main.py file to make reviewing easy. I fixed the header bar by moving it to the top.